Include a binary wrapper for the *latex binaries
pandoc doe snot like it when the *latex packages are batch files instead of exe files. So include a binary wrapper which does the same as the batch files.

Unfortunately, the wrapper is quite big because I was only able to build it in go :-) So only use it for the *latex binaries, not as a complete replacement for the batch files...